// client/state/Categories/categories.action-types.ts
export enum ActionTypes {
  CREATE_CATEGORY_START = "CREATE_CATEGORY_START",
  CREATE_CATEGORY_SUCCESS = "CREATE_CATEGORY_SUCCESS",
  CREATE_CATEGORY_ERROR = "CREATE_CATEGORY_ERROR",

  UPDATE_CATEGORY_START = "UPDATE_CATEGORY_START",
  UPDATE_CATEGORY_SUCCESS = "UPDATE_CATEGORY_SUCCESS",
  UPDATE_CATEGORY_ERROR = "UPDATE_CATEGORY_ERROR",

  DELETE_CATEGORY_START = "DELETE_CATEGORY_START",
  DELETE_CATEGORY_SUCCESS = "DELETE_CATEGORY_SUCCESS",
  DELETE_CATEGORY_ERROR = "DELETE_CATEGORY_ERROR",

  FETCH_CATEGORIES_START = "FETCH_CATEGORIES_START",
  FETCH_CATEGORIES_SUCCESS = "FETCH_CATEGORIES_SUCCESS",
  FETCH_CATEGORIES_ERROR = "FETCH_CATEGORIES_ERROR",
}